Feat: 보스유닛 공격 로직 구현

+++ src/main/java/com/stardy/poker_defense/unit/vo/Reward.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Reward {
     private Integer gold;
-    private Integer systemUnitId;
+    private Long systemUnitId;
 }
